requirements.txt
ShopifyAPI==12.2.0
requests==2.31.0 